+++ src/main/java/com/graphhopper/sna/graphcreators/GraphCreator.java
@@ -22,9 +22,25 @@
  * You should have received a copy of the GNU General Public License along with
  * GraphHopper-SNA. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.graphhopper.sna.model;
+package com.graphhopper.sna.graphcreators;
 
 import com.graphhopper.sna.data.IdInfo;
+import com.graphhopper.sna.model.DirectedG;
+import com.graphhopper.sna.model.DirectedG;
+import com.graphhopper.sna.model.DirectedPseudoG;
+import com.graphhopper.sna.model.DirectedPseudoG;
+import com.graphhopper.sna.model.DirectedWeightedPseudoG;
+import com.graphhopper.sna.model.DirectedWeightedPseudoG;
+import com.graphhopper.sna.model.Edge;
+import com.graphhopper.sna.model.Edge;
+import com.graphhopper.sna.model.KeyedGraph;
+import com.graphhopper.sna.model.KeyedGraph;
+import com.graphhopper.sna.model.PseudoG;
+import com.graphhopper.sna.model.PseudoG;
+import com.graphhopper.sna.model.UndirectedG;
+import com.graphhopper.sna.model.UndirectedG;
+import com.graphhopper.sna.model.WeightedPseudoG;
+import com.graphhopper.sna.model.WeightedPseudoG;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileNotFoundException;
